# Conflicts:
#	resources/views/emails/reminder.blade.php
#	routes/web.php

+++ database/migrations/2024_12_20_081324_create_sub_categories_table.php
@@ -13,10 +13,10 @@ return new class extends Migration
     {
         Schema::create('sub_categories', function (Blueprint $table) {
             $table->id();
-            $table->string('name');
-            $table->unsignedBigInteger('requirement')->default(0); // 非負の整数で、デフォルト値は 0
-            $table->foreignId('category_id')->constrained('categories')->cascadeOnDelete();
-            $table->timestamps();
+            $table->string('name'); // サブカテゴリ名
+            $table->decimal('requirement', 10, 5)->default(0.0); // 小数点対応のdecimal型に変更
+            $table->foreignId('category_id')->constrained('categories')->cascadeOnDelete(); // categoriesテーブルへの外部キー
+            $table->timestamps(); // created_at, updated_at
         });
     }
